test unsigned int to long'

+++ ft_printf/nbr_management/ft_uitoa_ext_buffer_sub_10.c
@@ -6,16 +6,16 @@
 /*   By: scarboni <scarboni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/26 08:00:02 by scarboni          #+#    #+#             */
-/*   Updated: 2020/08/15 14:49:12 by scarboni         ###   ########.fr       */
+/*   Updated: 2020/08/24 16:33:24 by scarboni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../ft_printf.h"
 
-int	ft_uitoa_ext_buffer_sub_10(unsigned int nbr, char *buffer, int base)
+int	ft_uitoa_ext_buffer_sub_10(unsigned long nbr, char *buffer, int base)
 {
-	unsigned int	i;
-	unsigned int	len;
+	unsigned long	i;
+	unsigned long	len;
 
 	if (base > 10 || !buffer)
 		return (-EXIT_FAILURE);
